requirements.txt
pandas
numpy
requests
python-dotenv
finance-datareader
duckduckgo-search
